Add files via upload
The real one I plan to name Story (this was just temp so I can have both later on. 
This is just on my own. This has the main method. Obviously the main method isn't permanent. It is just for my testing purposes. 

And it's not close to being done. Lots more to do.